academix/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chromedriver-autoinstaller==0.4.0
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
geckodriver-autoinstaller==0.1.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
pyspnego==0.8.0
pytz==2022.7.1
requests==2.28.2
requests-ntlm==1.2.0
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.14
wsproto==1.2.0
mechanize==0.4.8
